+++ src/imio/smartweb/core/webcomponents/src/components/News/ContactContent/ContactContent.jsx
@@ -1,19 +1,37 @@
-import { useHistory, useParams } from "react-router-dom";
+import { useHistory, useParams, useLocation } from "react-router-dom";
 import React, { useEffect, useState } from "react";
+import useAxios from "../../../hooks/useAxios";
+import useFilterQuery from "../../../hooks/useFilterQuery";
 
-const ContactContent = ({ contactArray, onChange }) => {
+const ContactContent = ({ queryUrl, onChange }) => {
     let history = useHistory();
-    const { id } = useParams();
+    const queryString = require("query-string");
+    const parsed = queryString.parse(useFilterQuery().toString());
+    const [params, setParams] = useState({});
+
     const [contactItem, setcontactItem] = useState({});
+    const { response, error, isLoading } = useAxios({
+        method: "get",
+        url: "",
+        baseURL: queryUrl,
+        headers: {
+            Accept: "application/json",
+        },
+        params: params,
+    });
+    useEffect(() => {
+        if (response !== null) {
+            setcontactItem(response.items[0]);
+        }
+    }, [response]);
 
     useEffect(() => {
-        if (contactArray.length > 0) {
-            var result = contactArray.filter((obj) => {
-                return obj.UID === id;
-            });
-            setcontactItem(result[0]);
-        }
-    }, [id, contactArray]);
+        setParams({
+            UID: parsed.u,
+            fullobjects: 1,
+        });
+    }, []);
+
 
     function handleClick() {
         history.push(".");
@@ -24,72 +42,8 @@ const ContactContent = ({ contactArray, onChange }) => {
             <button type="button" onClick={handleClick}>
                 Go home
             </button>
-            <div className="contactCard">
-                <div className="contactText">
-                    <div className="contactTextTitle">
-                        <span className="title">{contactItem.title}</span>
-                    </div>
-                    <div className="contactTextAll">
-                        {contactItem.category ? <span>{contactItem.category}</span> : ""}
-                        <div className="adresse">
-                            {contactItem.number ? <span>{contactItem.number + " "}</span> : ""}
-                            {contactItem.street ? <span>{contactItem.street + ", "}</span> : ""}
-                            {contactItem.complement ? (
-                                <span>{contactItem.complement + ", "}</span>
-                            ) : (
-                                ""
-                            )}
-                            {contactItem.zipcode ? <span>{contactItem.zipcode + " "}</span> : ""}
-                            {contactItem.city ? <span>{contactItem.city}</span> : ""}
-                        </div>
-                        <div className="itineraty">
-                            {contactItem.street ? (
-                                <a
-                                    href={
-                                        "https://www.google.com/maps/dir/?api=1&destination=" +
-                                        contactItem.street +
-                                        "+" +
-                                        contactItem.number +
-                                        "+" +
-                                        contactItem.complement +
-                                        "+" +
-                                        contactItem.zipcode +
-                                        "+" +
-                                        contactItem.city +
-                                        "+" +
-                                        contactItem.country
-                                    }
-                                >
-                                    Itinéraire
-                                </a>
-                            ) : (
-                                ""
-                            )}
-                        </div>
-                        <div className="phones">
-                            {contactItem.phones
-                                ? contactItem.phones.map((phone) => {
-                                      return <span>{phone.number}</span>;
-                                  })
-                                : ""}
-                        </div>
-                        <div className="mails">
-                            {contactItem.mails
-                                ? contactItem.mails.map((mail) => {
-                                      return <span>{mail.mail_address}</span>;
-                                  })
-                                : ""}
-                        </div>
-                        <div className="topics">
-                            {contactItem.topics
-                                ? contactItem.topics.map((mail) => {
-                                      return <span>{mail.title}</span>;
-                                  })
-                                : ""}
-                        </div>
-                    </div>
-                </div>
-            </div>
+            <span className="title">{contactItem.title}</span>
+
         </div>
     );
 };