+++ src/imio/smartweb/core/webcomponents/src/components/News/News.jsx
@@ -5,15 +5,25 @@ import Skeleton from "./Skeleton/Skeleton.jsx";
 import Filters from "./Filters/Filter";
 import ContactContent from "./ContactContent/ContactContent";
 import ContactList from "./ContactList/ContactList";
-import ContactMap from "./ContactMap/ContactMap";
 import useAxios from "../../hooks/useAxios";
 import "./News.scss";
+import useFilterQuery from "../../hooks/useFilterQuery";
 
-const Annuaire = (props) => {
+export default function News(props) {
+    return (
+        <Router>
+            <NewsView queryFilterUrl={props.queryFilterUrl} queryUrl={props.queryUrl} />
+        </Router>
+    );
+}
+
+const NewsView = (props) => {
+    const queryString = require("query-string");
+    const parsed = queryString.parse(useFilterQuery().toString());
     const [contactArray, setcontactArray] = useState([]);
     const [clickId, setClickId] = useState(null);
     const [params, setParams] = useState({});
-    const [filters, setFilters] = useState({});
+    const [filters, setFilters] = useState(parsed);
     const [batchSize, setBatchSize] = useState(5);
     const [refTop, setRefTop] = useState(null);
     const { response, error, isLoading } = useAxios(
@@ -56,19 +66,20 @@ const Annuaire = (props) => {
                 setRefTop(el.getBoundingClientRect().top);
             }}
             style={{ height: `calc(100vh -  ${refTop}px)` }}
-        >
-            {/* <h1>{"ddd"+props.queryUrl}</h1>
-            <h2>{"ddd"+props.queryFilterUrl}</h2> */}
+        >   
             <Router>
                 <div className="r-wrapper r-annuaire-wrapper">
                     <div className="r-result r-annuaire-result">
                         <Switch>
-                            <Route path={"/:id"}>
-                                <ContactContent onChange={clickID} contactArray={contactArray} />
+                            <Route path={"/:name"}>
+                                <ContactContent onChange={clickID} queryUrl={props.queryUrl} />
                             </Route>
                             <Route exact path="*">
                                 <div className="r-result-filter annuaire-result-filter">
-                                    <Filters url={props.queryFilterUrl} onChange={filtersChange} />
+                                    <Filters 
+                                    url={props.queryFilterUrl}
+                                    activeFilter={filters}
+                                    onChange={filtersChange} />
                                 </div>
                                 {isLoading ? (
                                     <div>
@@ -80,7 +91,6 @@ const Annuaire = (props) => {
                                         contactArray={contactArray}
                                         parentCallback={callback}
                                     />
-                                    // <h1>{props.queryUrl}</h1>
                                 )}
                             </Route>
                         </Switch>
@@ -91,26 +101,3 @@ const Annuaire = (props) => {
     );
 };
 
-// Thematiques == topi
-export default Annuaire;
-
-// const directoryApi = async () => {
-//   try {
-//     const contacts = await axios.get('https://annuaire.preprod.imio.be/@search?fullobjects=1', {
-//       headers: {
-//         "Accept": "application/json"
-//       }
-//     });
-//     setcontactArray(contacts.data.items);
-//     setcontactResults(contacts.data.items.length);
-//   } catch (error) {
-
-//   }
-// };
-
-// useEffect(() => {
-//   // setcontactArray(data.items);
-//   directoryApi();
-
-//   // eslint-disable-next-line
-// }, [contactResults, loadmore]);