+++ src/imio/smartweb/core/contents/rest/news/endpoint.py
@@ -44,6 +44,8 @@ class NewsFiltersEndpoint(BaseNewsEndpoint):
 
 class NewsEndpointGet(Service):
     def render(self):
+        response = self.request.response
+        response.setHeader("Content-Type", self.content_type)
         related_items = NewsEndpoint(self.context, self.request)
         return json.dumps(
             related_items(),
@@ -54,6 +56,8 @@ class NewsEndpointGet(Service):
 
 class NewsFiltersEndpointGet(Service):
     def render(self):
+        response = self.request.response
+        response.setHeader("Content-Type", self.content_type)
         related_items = NewsFiltersEndpoint(self.context, self.request)
         return json.dumps(
             related_items(),